requirements.txt
altgraph==0.17.4
amqp==5.3.1
asgiref==3.10.0
beautifulsoup4==4.14.2
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distlib==0.4.0
Django==5.2.7
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.0
fonttools==4.60.1
idna==3.11
kiwisolver==1.4.9
kombu==5.5.4
matplotlib==3.10.7
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pefile==2024.8.26
pillow==11.3.0
pip-review==1.3.0
pipenv==2025.0.4
platformdirs==4.5.0
prompt_toolkit==3.0.52
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
redis==6.4.0
requests==2.32.5
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
virtualenv==20.35.3
wcwidth==0.2.14
